refactor(CO-3211): update conversations type imports to use specific paths for better clarity
