+++ src/views/app/folder-panel/parts/row-info.tsx
@@ -10,7 +10,7 @@ import { Icon, Padding, Row, Text } from '@zextras/carbonio-design-system';
 import { Tag } from '@zextras/carbonio-ui-commons';
 
 import { getTimeLabel } from 'commons/utils';
-import { NormalizedConversation } from 'types/index.d';
+import { NormalizedConversation } from 'types/conversations';
 import { useTagExist } from 'ui-actions/tag-actions';
 
 type RowInfoProps = {
